Refactor imports to use new module structure for bean collection and component classes; add BeanCollection and Component classes with lifecycle management and error handling features.
