+++ py_spring_core/__init__.py
@@ -1,6 +1,6 @@
 from py_spring_core.core.application.py_spring_application import PySpringApplication
-from py_spring_core.core.entities.bean_collection import BeanCollection
-from py_spring_core.core.entities.component import Component, ComponentScope
+from py_spring_core.core.entities.bean_collection.bean_collection import BeanCollection
+from py_spring_core.core.entities.component.component import Component, ComponentScope
 from py_spring_core.core.entities.controllers.rest_controller import RestController
 from py_spring_core.core.entities.controllers.route_mapping import (
     DeleteMapping,
@@ -9,7 +9,7 @@ from py_spring_core.core.entities.controllers.route_mapping import (
     PostMapping,
     PutMapping,
 )
-from py_spring_core.core.entities.entity_provider import EntityProvider
+from py_spring_core.core.entities.entity_provider.entity_provider import EntityProvider
 from py_spring_core.core.entities.middlewares.middleware import Middleware
 from py_spring_core.core.entities.middlewares.middleware_registry import (
     MiddlewareRegistry,
