+++ py_spring_core/core/application/commons.py
@@ -1,5 +1,5 @@
-from py_spring_core.core.entities.bean_collection import BeanCollection
-from py_spring_core.core.entities.component import Component
+from py_spring_core.core.entities.bean_collection.bean_collection import BeanCollection
+from py_spring_core.core.entities.component.component import Component
 from py_spring_core.core.entities.controllers.rest_controller import RestController
 from py_spring_core.core.entities.properties.properties import Properties
