+++ py_spring_core/core/application/context/application_context.py
@@ -22,15 +22,15 @@ from py_spring_core.core.application.commons import AppEntities
 from py_spring_core.core.application.context.application_context_config import (
     ApplicationContextConfig,
 )
-from py_spring_core.core.entities.bean_collection import (
+from py_spring_core.core.entities.bean_collection.bean_collection import (
     BeanCollection,
     BeanConflictError,
     BeanView,
     InvalidBeanError,
 )
-from py_spring_core.core.entities.component import Component, ComponentScope
+from py_spring_core.core.entities.component.component import Component, ComponentScope
 from py_spring_core.core.entities.controllers.rest_controller import RestController
-from py_spring_core.core.entities.entity_provider import EntityProvider
+from py_spring_core.core.entities.entity_provider.entity_provider import EntityProvider
 from py_spring_core.core.entities.properties.properties import Properties
 from py_spring_core.core.entities.properties.properties_loader import _PropertiesLoader
