+++ py_spring_core/core/entities/component/component.py
