+++ py_spring_core/core/entities/entity_provider/entity_provider.py
@@ -2,8 +2,8 @@ from dataclasses import dataclass, field
 from typing import Any, Optional, Type
 
 from py_spring_core.core.application.commons import AppEntities
-from py_spring_core.core.entities.bean_collection import BeanCollection
-from py_spring_core.core.entities.component import Component
+from py_spring_core.core.entities.bean_collection.bean_collection import BeanCollection
+from py_spring_core.core.entities.component.component import Component
 from py_spring_core.core.entities.controllers.rest_controller import RestController
 from py_spring_core.core.entities.properties.properties import Properties
